Modify ae353_drone.py to allow the user to selectively print results.
I added an addition argument to the run() function to allow for the user to choose if they want to print results or not. If the user does not specify, it will print by default. Otherwise you can use run(print_results=True) to print or run(print_results=False) to not print.